REFACTOR: make providers folder
