+++ src/ai_code_review/providers/lmstudio_provider.py
@@ -1,4 +1,4 @@
-"""OllamaProvider class for interacting with the Ollama API."""
+"""LMStudioProvider class for interacting with the LMStudio API."""
 
 from collections.abc import AsyncIterator
 
@@ -6,16 +6,16 @@ import aiohttp
 import orjson as json
 from loguru import logger
 
-from ai_code_review.base_llm_provider import BaseLLMProvider
+from ai_code_review.providers.base_llm_provider import BaseLLMProvider
 
 HTTP_OK_STATUS = 200
 
 
-class OllamaProvider(BaseLLMProvider):
-    """Provider class for interacting with the Ollama API.
+class LMStudioProvider(BaseLLMProvider):
+    """Provider class for interacting with the LMStudio API.
 
-    This class handles the connection to the Ollama server and provides methods
-    to stream responses from the Ollama model.
+    This class handles the connection to the LMStudio server and provides methods
+    to stream responses from the LMStudio model.
     """
 
     def __init__(
@@ -25,8 +25,8 @@ class OllamaProvider(BaseLLMProvider):
         model: str,
         auth_token: str | None = None,
     ):
-        """Initialize the OllamaProvider with the server URL, port, model, and optional auth token."""
-        self._url = f"http://{url}:{port}/api/generate"
+        """Initialize the LMStudioProvider with the server URL, port, model, and optional auth token."""
+        self._url = f"http://{url}:{port}/v1/chat/completions"
         self._model = model
         self._headers = {"Content-Type": "application/json"}
         if auth_token is not None:
@@ -37,23 +37,26 @@ class OllamaProvider(BaseLLMProvider):
         prompt: str,
         system_prompt: str | None = None,
     ) -> AsyncIterator[str]:
-        """Stream a response from the Ollama model.
+        """Stream a response from the LMStudio model.
 
         Args:
             prompt (str): The prompt to send to the model.
             system_prompt (str | None): An optional system prompt for the model.
 
         Returns:
-            AsyncIterator[str]: The streamed response from the Ollama server.
+            AsyncIterator[str]: The streamed response from the LMStudio server.
 
         """
+        messages = []
+        if system_prompt:
+            messages.append({"role": "system", "content": system_prompt})
+        messages.append({"role": "user", "content": prompt})
+
         payload = {
             "model": self._model,
-            "prompt": prompt,
+            "messages": messages,
             "stream": True,
         }
-        if system_prompt:
-            payload["system"] = system_prompt
 
         # Mask sensitive header values before logging
         safe_headers = self._headers.copy()
@@ -68,14 +71,21 @@ class OllamaProvider(BaseLLMProvider):
             if response.status != HTTP_OK_STATUS:
                 response_text = await response.text()
                 logger.error(f"Error: {response.status} - Response: {response_text}")
-                raise RuntimeError(f"Ollama API returned status {response.status}: {response_text}")
+                raise RuntimeError(f"LMStudio API returned status {response.status}: {response_text}")
 
             async for line in response.content:
-                line_str = line.decode("utf-8")
-                if line_str.strip():
+                line_str = line.decode("utf-8").strip()
+                if line_str and line_str.startswith("data: "):
+                    data_str = line_str[6:]  # Remove "data: " prefix
+                    if data_str == "[DONE]":
+                        break
                     try:
-                        data = json.loads(line_str)
-                        yield data.get("response", "")
+                        data = json.loads(data_str)
+                        if "choices" in data and data["choices"]:
+                            delta = data["choices"][0].get("delta", {})
+                            content = delta.get("content", "")
+                            if content:
+                                yield content
                     except json.JSONDecodeError as e:
                         logger.error(f"JSON decode error: {e}")
-                        break
+                        continue
