+++ src/ai_code_review/providers/ollama_provider.py
@@ -1,4 +1,4 @@
-"""LMStudioProvider class for interacting with the LMStudio API."""
+"""OllamaProvider class for interacting with the Ollama API."""
 
 from collections.abc import AsyncIterator
 
@@ -6,16 +6,16 @@ import aiohttp
 import orjson as json
 from loguru import logger
 
-from ai_code_review.base_llm_provider import BaseLLMProvider
+from ai_code_review.providers.base_llm_provider import BaseLLMProvider
 
 HTTP_OK_STATUS = 200
 
 
-class LMStudioProvider(BaseLLMProvider):
-    """Provider class for interacting with the LMStudio API.
+class OllamaProvider(BaseLLMProvider):
+    """Provider class for interacting with the Ollama API.
 
-    This class handles the connection to the LMStudio server and provides methods
-    to stream responses from the LMStudio model.
+    This class handles the connection to the Ollama server and provides methods
+    to stream responses from the Ollama model.
     """
 
     def __init__(
@@ -25,8 +25,8 @@ class LMStudioProvider(BaseLLMProvider):
         model: str,
         auth_token: str | None = None,
     ):
-        """Initialize the LMStudioProvider with the server URL, port, model, and optional auth token."""
-        self._url = f"http://{url}:{port}/v1/chat/completions"
+        """Initialize the OllamaProvider with the server URL, port, model, and optional auth token."""
+        self._url = f"http://{url}:{port}/api/generate"
         self._model = model
         self._headers = {"Content-Type": "application/json"}
         if auth_token is not None:
@@ -37,26 +37,23 @@ class LMStudioProvider(BaseLLMProvider):
         prompt: str,
         system_prompt: str | None = None,
     ) -> AsyncIterator[str]:
-        """Stream a response from the LMStudio model.
+        """Stream a response from the Ollama model.
 
         Args:
             prompt (str): The prompt to send to the model.
             system_prompt (str | None): An optional system prompt for the model.
 
         Returns:
-            AsyncIterator[str]: The streamed response from the LMStudio server.
+            AsyncIterator[str]: The streamed response from the Ollama server.
 
         """
-        messages = []
-        if system_prompt:
-            messages.append({"role": "system", "content": system_prompt})
-        messages.append({"role": "user", "content": prompt})
-
         payload = {
             "model": self._model,
-            "messages": messages,
+            "prompt": prompt,
             "stream": True,
         }
+        if system_prompt:
+            payload["system"] = system_prompt
 
         # Mask sensitive header values before logging
         safe_headers = self._headers.copy()
@@ -71,21 +68,14 @@ class LMStudioProvider(BaseLLMProvider):
             if response.status != HTTP_OK_STATUS:
                 response_text = await response.text()
                 logger.error(f"Error: {response.status} - Response: {response_text}")
-                raise RuntimeError(f"LMStudio API returned status {response.status}: {response_text}")
+                raise RuntimeError(f"Ollama API returned status {response.status}: {response_text}")
 
             async for line in response.content:
-                line_str = line.decode("utf-8").strip()
-                if line_str and line_str.startswith("data: "):
-                    data_str = line_str[6:]  # Remove "data: " prefix
-                    if data_str == "[DONE]":
-                        break
+                line_str = line.decode("utf-8")
+                if line_str.strip():
                     try:
-                        data = json.loads(data_str)
-                        if "choices" in data and data["choices"]:
-                            delta = data["choices"][0].get("delta", {})
-                            content = delta.get("content", "")
-                            if content:
-                                yield content
+                        data = json.loads(line_str)
+                        yield data.get("response", "")
                     except json.JSONDecodeError as e:
                         logger.error(f"JSON decode error: {e}")
-                        continue
+                        break
